Add bkg estimation, add detrending, add more error handling.
Added photutils background estimation preprocessing step to
line detection for those low quality images.

Added detrending to line profile fitting for cases where
the edge of the image appears in the rotated and cropped
image. This required updating the fit function in
gaussian. To maintain compatibility it required adoption
of a function signature that isn't the best.
TODO: change fit function signature to something better

Added some error checks:
    - fixed the cropping coordinate determination. When
      line was parallel to an image edge the crop would
      return an empy image because one of the indices
      would go out of bound.
    - added a check to skip further processing when no
      lines were found by the line detector
    - when line detector only finds 1 line the shape of
      the returned lines array is different than when
      many are returned.

Fixed an bug in transform_rho_theta with the passing
of cart_coordinates, expected a tuple, got a list of
tuples. This is important in edgecases when the line
is vertical or horizontal.

Added a bunch of debug plots that need to be removed
or protected by an if.